// src/layout/types.ts
export enum AlignItems {
  Start = "flex-start",
  End = "flex-end",
  Center = "center",
  Baseline = "baseline",
  Stretch = "stretch",
}
// ...
export enum Display {
  Block = "block",
  Flex = "flex",
  Initial = "initial",
  Inline = "inline",
  InlineBlock = "inline-block",
  InlineFlex = "inline-flex",
  Hide = "none",
  HideAccessible = "hide-accessible",
}

export type DisplayValue =
  | "block"
  | "flex"
  | "initial"
  | "inline"
  | "inline-block"
  | "inline-flex"
  | "none"
  | "hide-accessible"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum FlexDirection {
  Column = "column",
  ColumnReverse = "column-reverse",
  Row = "row",
  RowReverse = "row-reverse",
}

export type FlexDirectionValue =
  | "column"
  | "column-reverse"
  | "row"
  | "row-reverse"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum FlexWrap {
  Wrap = "wrap",
  NoWrap = "nowrap",
  WrapReverse = "wrap-reverse",
}
// ...
export enum Position {
  Relative = "relative",
  Absolute = "absolute",
  Fixed = "fixed",
}
// ...
export enum Background {
  Base = "base",
  Alt = "alt",
  Alt2 = "alt-2",
  Overlay = "overlay",
  Accent = "accent",
  AccentAlt = "accent-alt",
  AccentAlt2 = "accent-alt-2",
  Inherit = "inherit",
}

export type BackgroundValue =
  | "base"
  | "alt"
  | "alt-2"
  | "overlay"
  | "accent"
  | "accent-alt"
  | "accent-alt-2"
  | "inherit"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum Color {
  Base = "base",
  Alt = "alt",
  Alt2 = "alt-2",
  Link = "link",
  Live = "live",
  Error = "error",
  Overlay = "overlay",
  OverlayAlt = "overlay-alt",
  Inherit = "inherit",
}

export type ColorValue =
  | "base"
  | "alt"
  | "alt-2"
  | "link"
  | "live"
  | "error"
  | "overlay"
  | "overlay-alt"
  | "inherit"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export type Elevation = 0 | 1 | 2 | 3 | 4 | 5;